Correct the fact that all access checking were wrongly factorized

+++ src/Controller/admin/AdminFeedbackController.php
@@ -23,7 +23,6 @@ class AdminFeedbackController extends AbstractController
     {
         $this->paginator = $paginator;
         $this->repo = $repo;
-        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Veuillez vous connecter.");
     }
 
     /**
@@ -33,6 +32,7 @@ class AdminFeedbackController extends AbstractController
      */
     public function index(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Veuillez vous connecter.");
         $feedbacks = $this->paginator->paginate($this->repo->findAll(), $request->query->getInt('page', 1), 11);
 
         return $this->render('admin/feedback/index.html.twig', [
@@ -49,6 +49,7 @@ class AdminFeedbackController extends AbstractController
      */
     public function delete(Feedback $feedback, EntityManagerInterface $em): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Veuillez vous connecter.");
         $em->remove($feedback);
         $em->flush();
         $this->addFlash('success', 'Ce commentaire a été supprimé avec succès.');
@@ -62,6 +63,7 @@ class AdminFeedbackController extends AbstractController
      */
     public function sort(string $orderBy, string $order, Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Veuillez vous connecter.");
         $feedbacks = $this->paginator->paginate($this->repo->orderFeedbacks($orderBy, $order), $request->query->getInt('page', 1), 11);
         return $this->render('admin/feedback/index.html.twig', [
             'feedbacks' => $feedbacks,
