+++ src/Controller/admin/AdminTripController.php
@@ -27,8 +27,6 @@ class AdminTripController extends AbstractController
         $this->tripRepository = $tripRepository;
         $this->paginator = $paginatorInterface;
         $this->em = $em;
-        // Check roles
-        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Veuillez vous connecter.");
     }
     /**
      * Show all trips
@@ -36,7 +34,8 @@ class AdminTripController extends AbstractController
      * @return Response
      */
     public function index(Request $request): Response
-    {
+    { // Check roles
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Veuillez vous connecter.");
         $trips = $this->paginator->paginate($this->tripRepository->findBy(['reserved' => false]), $request->query->getInt('page', 1), 11);
         return $this->render('admin/trip/index.html.twig', [
             'trips' => $trips,
@@ -50,7 +49,8 @@ class AdminTripController extends AbstractController
      * @return Response
      */
     public function create(Request $request): Response
-    {
+    { // Check roles
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Veuillez vous connecter.");
         $form = $this->createForm(TripType::class);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -75,7 +75,8 @@ class AdminTripController extends AbstractController
      * @return Response
      */
     public function delete(Trip $trip): Response
-    {
+    { // Check roles
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Veuillez vous connecter.");
         $this->em->remove($trip);
         $this->em->flush();
         $this->addFlash('success', 'La suppression du voyage ' . $trip->getName() . ' a été effectué avec succès.');
@@ -87,7 +88,8 @@ class AdminTripController extends AbstractController
      * @return Response
      */
     public function edit(Trip $trip, Request $request): Response
-    {
+    { // Check roles
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Veuillez vous connecter.");
         $reserved = $trip->getReserved();
         $form = $this->createForm(TripType::class, $trip);
         $form->handleRequest($request);
@@ -111,7 +113,8 @@ class AdminTripController extends AbstractController
      * @return Response
      */
     public function sort(Request $request, string $orderBy, string $order): Response
-    {
+    { // Check roles
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Veuillez vous connecter.");
         $trips = $this->paginator->paginate($this->tripRepository->orderTrips($orderBy, $order), $request->query->getInt('page', 1), 11);
         return $this->render('admin/trip/index.html.twig', [
             'trips' => $trips,
@@ -127,7 +130,8 @@ class AdminTripController extends AbstractController
      * @return Response
      */
     public function resIndex(Request $request): Response
-    {
+    { // Check roles
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Veuillez vous connecter.");
         $trips = $this->paginator->paginate($this->tripRepository->findBy(['reserved' => true]), $request->query->getInt('page', 1), 11);
         return $this->render('admin/trip/index.html.twig', [
             'trips' => $trips,
